requirements.txt
fastapi
uvicorn
ultralytics
numpy
pillow
